add compoundSelectors to redux
