+++ src/state/selection/selectors/basic.ts
@@ -13,3 +13,5 @@ export const getSelectedAgentMetadata = (state: State) =>
     state.selection.selectedAgentMetadata;
 export const getSelectedUIDisplayData = (state: State) =>
     state.selection.selectedUIDisplayData;
+export const getCurrentColorSettings = (state: State) =>
+    state.selection.currentColorSettings;
