+++ src/state/trajectory/selectors/index.ts
@@ -1,12 +1,11 @@
 import { createSelector } from "reselect";
-import { UIDisplayData } from "@aics/simularium-viewer";
 
 import {
     isNetworkSimFileInterface,
     LocalSimFile,
     NetworkedSimFile,
 } from "../types";
-import { getSimulariumFile, getDefaultUIDisplayData } from "./basic";
+import { getSimulariumFile } from "./basic";
 
 export const getIsNetworkedFile = createSelector(
     [getSimulariumFile],
@@ -15,29 +14,6 @@ export const getIsNetworkedFile = createSelector(
             return false;
         }
         return isNetworkSimFileInterface(simFile);
-    }
-);
-
-export const getUiDisplayDataTree = createSelector(
-    [getDefaultUIDisplayData],
-    (uiDisplayData: UIDisplayData) => {
-        if (!uiDisplayData.length) {
-            return [];
-        }
-        return uiDisplayData.map((agent) => ({
-            title: agent.name,
-            key: agent.name,
-            color: agent.color,
-            children: agent.displayStates.length
-                ? [
-                      ...agent.displayStates.map((state) => ({
-                          label: state.name,
-                          value: state.id,
-                          color: state.color,
-                      })),
-                  ]
-                : [],
-        }));
     }
 );
